refactor(codec): restructure codec directory to vendor/endpoint layout
Reorganize codec subdirectories from flat protocol-level grouping to
a two-level vendor/endpoint hierarchy:
- anthropic_messages/ → anthropic/messages/
- google_generative/ → google/gemini/
- openai_compatible/ → openai/compatible/
- openai_responses/ → openai/responses/

Add vendor-level mod.rs files (anthropic/, google/, openai/) and
update all `use` paths across codec, dispatcher, ingress, and tests.
No behavioral changes — pure directory restructuring.
