+++ crates/nyro-core/src/protocol/codec/anthropic/messages/decoder.rs
