+++ crates/nyro-core/src/protocol/codec/anthropic/messages/encoder.rs
