+++ crates/nyro-core/src/protocol/codec/google/gemini/decoder.rs
