+++ crates/nyro-core/src/protocol/codec/google/gemini/encoder.rs
