+++ crates/nyro-core/src/protocol/codec/google/gemini/generate_content.rs
