+++ crates/nyro-core/src/protocol/codec/google/gemini/mod.rs
