+++ crates/nyro-core/src/protocol/codec/google/gemini/stream.rs
