+++ crates/nyro-core/src/protocol/codec/google/gemini/types.rs
