+++ crates/nyro-core/src/protocol/codec/google/mod.rs
@@ -1,0 +1,1 @@
+pub mod gemini;
