+++ crates/nyro-core/src/protocol/codec/mod.rs
@@ -2,9 +2,8 @@
 //! request/response/stream codecs **and** the thin `EndpointHandler`
 //! registration shell for every endpoint.
 
-pub mod anthropic_messages;
-pub mod google_generative;
-pub mod openai_compatible;
-pub mod openai_responses;
+pub mod anthropic;
+pub mod google;
+pub mod openai;
 pub mod reasoning;
 pub mod tool_correlation;
