+++ crates/nyro-core/src/protocol/codec/openai/compatible/chat_completions.rs
