+++ crates/nyro-core/src/protocol/codec/openai/compatible/decoder.rs
