+++ crates/nyro-core/src/protocol/codec/openai/compatible/embeddings.rs
