+++ crates/nyro-core/src/protocol/codec/openai/compatible/mod.rs
