+++ crates/nyro-core/src/protocol/codec/openai/compatible/stream.rs
