+++ crates/nyro-core/src/protocol/codec/openai/compatible/types.rs
