+++ crates/nyro-core/src/protocol/codec/openai/mod.rs
@@ -1,0 +1,2 @@
+pub mod compatible;
+pub mod responses;
