+++ crates/nyro-core/src/protocol/codec/openai/responses/decoder.rs
