+++ crates/nyro-core/src/protocol/codec/openai/responses/encoder.rs
