+++ crates/nyro-core/src/protocol/codec/openai/responses/formatter.rs
