+++ crates/nyro-core/src/protocol/codec/openai/responses/mod.rs
