+++ crates/nyro-core/src/protocol/codec/openai/responses/parser.rs
