+++ crates/nyro-core/src/protocol/codec/openai/responses/responses.rs
