+++ crates/nyro-core/src/protocol/codec/openai/responses/stream.rs
